Add all sub-patterns detector

+++ src/gcdr.cpp
@@ -5,7 +5,7 @@
 
 size_t Graph::cw_in(size_t v) const {
     size_t ret = 1;
-    for (int i = 0; i < size(); ++i) {
+    for (size_t i = 0; i < size(); ++i) {
         ret *= edge(i, v);
     }
     return ret;
@@ -13,7 +13,7 @@ size_t Graph::cw_in(size_t v) const {
 
 size_t Graph::cw_out(size_t v) const {
     size_t ret = 1;
-    for (int i = 0; i < size(); ++i) {
+    for (size_t i = 0; i < size(); ++i) {
         ret *= edge(v, i);
     }
     return ret;
@@ -21,8 +21,8 @@ size_t Graph::cw_out(size_t v) const {
 
 void Graph::print_gcdr() const {
     auto n = size();
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < n; ++j) {
             std::cout << nodes[i].name << " --" << edge(i, j) << "--> "
                       << nodes[j].name << std::endl;
         }
